backend/requirements.txt
# Core Web
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Configuration
python-dotenv>=1.0.0

# LangChain & RAG
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.3.1


# Vector Database & Embeddings
chromadb>=0.4.22
pypika>=0.48.9
hnswlib>=0.7.0
posthog>=2.4.2
tokenizers>=0.15.1

# Database
mysql-connector-python>=8.2.0
sqlalchemy>=2.0.25
pymysql>=1.1.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.3

# LLM & AI
transformers>=4.36.2
torch>=2.1.2
mistralai>=0.0.11
huggingface-hub>=0.20.2

# Documents
PyPDF2>=3.0.1

# Text-to-Speech
gtts>=2.4.0
deep-translator>=1.11.4

# Utilities
requests>=2.31.0
urllib3>=1.26.18
